+++ src/HEADERS/WardIdentity.h
@@ -7,7 +7,7 @@
 #define WARDIDENTITY_H
 
 /**
-   @fn void compute_p_psq( double **p , double *psq , const struct veclist *list , const int NMOM )
+   @fn void compute_p_psq( double **p , double *psq , const struct veclist *list , const size_t NMOM )
    @brief computes the momenta
    @warning only 2sin(q/2) momentum definition used at the moment
  */
@@ -15,20 +15,20 @@ void
 compute_p_psq( double **p ,
 	       double *psq ,
 	       const struct veclist *list ,
-	       const int NMOM ) ;
+	       const size_t NMOM ) ;
 
 /**
-   @fn void compute_WI( const struct PIdata *data , const double **p , const struct veclist *list , const int NMOM )
+   @fn void compute_WI( const struct PIdata *data , const double **p , const struct veclist *list , const size_t NMOM )
    @brief print to stdout the momentum space WIs
  */
 void
 compute_WI( const struct PIdata *data ,
 	    const double **p ,
 	    const struct veclist *list ,
-	    const int NMOM) ;
+	    const size_t NMOM) ;
 
 /**
-   @fn void correct_WI( struct PIdata *data , const correction_dir corr_dir , const struct veclist *list , const int NMOM )
+   @fn void correct_WI( struct PIdata *data , const correction_dir corr_dir , const struct veclist *list , const size_t NMOM )
    @brief perform WI correction on idx
    @warning overwrites data
  */
@@ -36,7 +36,7 @@ void
 correct_WI( struct PIdata *data ,
 	    const correction_dir corr_dir ,
 	    const struct veclist *list ,
-	    const int NMOM ) ;
+	    const size_t NMOM ) ;
 
 /**
    @fn void WI_configspace_bwd( const struct PIdata *data , const struct site *lat )
