+++ src/HEADERS/wrap_baryons.h
@@ -6,7 +6,7 @@
 #define WRAP_BARYONS_H
 
 /**
-   @fn int contract_baryons( struct propagator *prop , const struct baryon_info *baryons , const struct cut_info CUTINFO , const int nbaryons )
+   @fn int contract_baryons( struct propagator *prop , const struct baryon_info *baryons , const struct cut_info CUTINFO , const size_t nbaryons )
    @brief baryon contraction logic
    @return #SUCCESS or #FAILURE
  */
@@ -14,6 +14,6 @@ int
 contract_baryons( struct propagator *prop ,
 		  const struct baryon_info *baryons ,
 		  const struct cut_info CUTINFO , 
-		  const int nbaryons ) ;
+		  const size_t nbaryons ) ;
 
 #endif
